requirements.txt
asyncio
discord.py
googletrans